move existing updateProfile tests
